Take the float narrowing out of the HTTP-date branch
`int(time.Until(t).Seconds())` narrows a float64, and Go leaves an
out-of-range float-to-int conversion implementation-defined. Where int is
32 bits, a date beyond ~68 years out — and time.Until saturates at ~292
for the year-9999 dates a server may legally send — produced a value that
read as non-positive, so the header was discarded and the loop fell back
to its millisecond backoff. That is the outcome the clamp exists to
prevent, reached by the one path the clamp ran too late to guard.

Whole seconds now come from integer division in the Duration domain,
which cannot leave int64, and clampRetryAfterSeconds takes an int64 and
bounds it by int's range as well as the Duration's. Cannot be executed on
a 64-bit host, so the new test pins the resulting contract rather than
claiming a red proof; GOARCH=386 and arm vet clean.

The probe's 5s hang guard could not fire either: Get was called inline, so
an uninterruptible wait never reached the elapsed check and the package
timeout reported it instead — most visibly now that an over-range delay
saturates at 292 years rather than wrapping negative. Get moves to a
goroutine behind a select. Proved by making the retry wait a plain
time.Sleep: the guard now fails in 5.00s naming the delay
(2562047h47m16s) instead of hanging.

SPEC said "Go saturates this conversion" where only the hand-written
client does; the generated loop keeps its own unclamped copy. Says so now,
pointing at #798.

+++ go/pkg/basecamp/errors.go
@@ -199,7 +199,7 @@ func ErrRateLimit(retryAfter int) *Error {
 	// whose seconds→Duration conversion overflows would make the retry loop
 	// wait a negative delay, i.e. not wait at all. The hint already treated
 	// non-positive as absent; now the field agrees with it.
-	retryAfter = clampRetryAfterSeconds(retryAfter)
+	retryAfter = clampRetryAfterSeconds(int64(retryAfter))
 
 	hint := "Try again later"
 	if retryAfter > 0 {
